+++ resources/views/manager/edit.blade.php
@@ -14,6 +14,7 @@
         <x-chief::links.window />
 
         <x-chief::field.window tagged="sidebar" />
+        <x-chief::field.window title="Seo" tagged="seo" />
     </x-slot>
 
 </x-chief::page>