+++ src/Fragments/Fragmentable.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Thinktomorrow\Chief\Fragments;
 
+use Illuminate\Contracts\View\View;
 use Thinktomorrow\Chief\Fragments\Database\FragmentModel;
 use Thinktomorrow\Chief\ManagedModels\ManagedModel;
 use Thinktomorrow\Chief\Shared\Concerns\Viewable\ViewableContract;
@@ -13,7 +14,7 @@ interface Fragmentable extends ManagedModel, ViewableContract
 
     public function renderFragment($owner, $loop, $viewData = []): string;
 
-    public function renderForm(): string;
+    public function adminView(): View;
 
     public function setFragmentModel(FragmentModel $fragmentModel): self;